backend/requirements.txt
# Hygieia Backend Requirements
# ============================

# Core Flask Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# Authentication
Werkzeug>=3.0.0
PyJWT>=2.8.0

# Environment
python-dotenv>=1.0.0

# AI/ML Models
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0

# TensorFlow for skin diagnosis model
tensorflow>=2.15.0

# Image Processing
Pillow>=10.0.0

# OpenAI for Dr. Hygieia Chat
openai>=1.0.0

# Utilities
requests>=2.31.0

lightgbm
xgboost
pandas
colorama
opencv-python
gunicorn
psycopg2
tqdm
